+++ src/main/java/game/JanggiGame.java
@@ -1,4 +1,4 @@
-package domain.game;
+package game;
 
 import domain.Position;
 import domain.Team;
